fix(p1): keep new identifiers JSON-safe and schema-consistent

+++ backend/prisma/migrations/20260812103000_p1_revenue_core/migration.sql
@@ -2,7 +2,7 @@
 -- This migration intentionally reuses existing customers, outlets, items, transactions and accounting ledgers.
 
 CREATE TABLE IF NOT EXISTS public.crm_opportunities (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   tenant_id INTEGER NOT NULL,
   outlet_id INTEGER,
   customer_id INTEGER,
@@ -30,9 +30,9 @@ CREATE INDEX IF NOT EXISTS idx_crm_opportunities_owner
   ON public.crm_opportunities (tenant_id, owner_user_id);
 
 CREATE TABLE IF NOT EXISTS public.crm_activities (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   tenant_id INTEGER NOT NULL,
-  opportunity_id BIGINT NOT NULL REFERENCES public.crm_opportunities(id) ON DELETE CASCADE,
+  opportunity_id INTEGER NOT NULL REFERENCES public.crm_opportunities(id) ON DELETE CASCADE,
   activity_type VARCHAR(30) NOT NULL DEFAULT 'follow_up',
   summary VARCHAR(255) NOT NULL,
   due_at TIMESTAMPTZ,
@@ -49,7 +49,7 @@ CREATE INDEX IF NOT EXISTS idx_crm_activities_tenant_due
 
 CREATE SEQUENCE IF NOT EXISTS public.sales_quotation_number_seq START 1;
 CREATE TABLE IF NOT EXISTS public.sales_quotations (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   tenant_id INTEGER NOT NULL,
   outlet_id INTEGER,
   customer_id INTEGER,
@@ -62,7 +62,7 @@ CREATE TABLE IF NOT EXISTS public.sales_quotations (
   tax_amount NUMERIC(15,2) NOT NULL DEFAULT 0,
   total NUMERIC(15,2) NOT NULL DEFAULT 0,
   notes TEXT,
-  opportunity_id BIGINT REFERENCES public.crm_opportunities(id) ON DELETE SET NULL,
+  opportunity_id INTEGER REFERENCES public.crm_opportunities(id) ON DELETE SET NULL,
   created_by INTEGER,
   created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
   updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
@@ -77,8 +77,8 @@ CREATE INDEX IF NOT EXISTS idx_sales_quotations_customer
   ON public.sales_quotations (tenant_id, customer_id);
 
 CREATE TABLE IF NOT EXISTS public.sales_quotation_items (
-  id BIGSERIAL PRIMARY KEY,
-  quotation_id BIGINT NOT NULL REFERENCES public.sales_quotations(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  quotation_id INTEGER NOT NULL REFERENCES public.sales_quotations(id) ON DELETE CASCADE,
   item_id INTEGER,
   description VARCHAR(255) NOT NULL,
   quantity NUMERIC(12,3) NOT NULL,
@@ -92,11 +92,11 @@ CREATE TABLE IF NOT EXISTS public.sales_quotation_items (
 
 CREATE SEQUENCE IF NOT EXISTS public.sales_order_number_seq START 1;
 CREATE TABLE IF NOT EXISTS public.sales_orders (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   tenant_id INTEGER NOT NULL,
   outlet_id INTEGER,
   customer_id INTEGER,
-  quotation_id BIGINT UNIQUE REFERENCES public.sales_quotations(id) ON DELETE SET NULL,
+  quotation_id INTEGER UNIQUE REFERENCES public.sales_quotations(id) ON DELETE SET NULL,
   sales_order_number VARCHAR(80) NOT NULL,
   status VARCHAR(30) NOT NULL DEFAULT 'confirmed',
   currency VARCHAR(3) NOT NULL DEFAULT 'IDR',
@@ -119,8 +119,8 @@ CREATE INDEX IF NOT EXISTS idx_sales_orders_customer
   ON public.sales_orders (tenant_id, customer_id);
 
 CREATE TABLE IF NOT EXISTS public.sales_order_items (
-  id BIGSERIAL PRIMARY KEY,
-  sales_order_id BIGINT NOT NULL REFERENCES public.sales_orders(id) ON DELETE CASCADE,
+  id SERIAL PRIMARY KEY,
+  sales_order_id INTEGER NOT NULL REFERENCES public.sales_orders(id) ON DELETE CASCADE,
   item_id INTEGER,
   description VARCHAR(255) NOT NULL,
   quantity NUMERIC(12,3) NOT NULL,
@@ -131,25 +131,25 @@ CREATE TABLE IF NOT EXISTS public.sales_order_items (
 );
 
 CREATE TABLE IF NOT EXISTS public.loyalty_wallets (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   tenant_id INTEGER NOT NULL,
   customer_id INTEGER NOT NULL,
-  points_balance BIGINT NOT NULL DEFAULT 0,
+  points_balance INTEGER NOT NULL DEFAULT 0,
   monetary_balance NUMERIC(15,2) NOT NULL DEFAULT 0,
-  lifetime_points_earned BIGINT NOT NULL DEFAULT 0,
-  lifetime_points_redeemed BIGINT NOT NULL DEFAULT 0,
+  lifetime_points_earned INTEGER NOT NULL DEFAULT 0,
+  lifetime_points_redeemed INTEGER NOT NULL DEFAULT 0,
   created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
   updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
   CONSTRAINT loyalty_wallet_unique UNIQUE (tenant_id, customer_id)
 );
 
 CREATE TABLE IF NOT EXISTS public.loyalty_ledger (
-  id BIGSERIAL PRIMARY KEY,
+  id SERIAL PRIMARY KEY,
   tenant_id INTEGER NOT NULL,
   customer_id INTEGER NOT NULL,
-  wallet_id BIGINT NOT NULL REFERENCES public.loyalty_wallets(id) ON DELETE CASCADE,
+  wallet_id INTEGER NOT NULL REFERENCES public.loyalty_wallets(id) ON DELETE CASCADE,
   entry_type VARCHAR(20) NOT NULL,
-  points_delta BIGINT NOT NULL DEFAULT 0,
+  points_delta INTEGER NOT NULL DEFAULT 0,
   monetary_delta NUMERIC(15,2) NOT NULL DEFAULT 0,
   reference_type VARCHAR(50),
   reference_id VARCHAR(100),
